아이템 및 좀비 기능 추가 및 수정
- 좀비 및 아이템 스크럽터블 오브젝트로 변경 및 좀비에게 랜덤 아이템 생성

+++ Assets/Script/WeaponController.cs
@@ -15,7 +15,7 @@ public class WeaponController : MonoBehaviourPunCallbacks
     [Tooltip("총 발사 사거리")]
     public float range = 100f;
     [Tooltip("한 발당 데미지")]
-    public float damage = 50f;
+    public int damage = 50;
 
     public int currentAmmo = 7;
     public int remaining_Ammo = 30; // 남은 탄약 개수
